refactor: update toast imports and remove unused constants
