+++ src/utils/constants.ts
@@ -1,5 +1,3 @@
-import { ToastOptions } from 'react-toastify';
-
 export const CODE_VERIFIER_LOCAL_STORAGE_NAME = 'a12n_ADMIN_UI_CODE_VERIFIER';
 export const AUTHORIZATION_CODE_QUERY_PARAM_NAME = 'code';
 export const POST_AUTH_REDIRECT_QUERY_PARAM_NAME = 'redirect';
@@ -42,33 +40,3 @@ export enum HTTP_METHODS {
     PATCH = 'patch',
     DELETE = 'delete',
 }
-
-export const DEFAULT_TOAST_OPTIONS: ToastOptions = {
-    position: 'top-right',
-    autoClose: 3000,
-    hideProgressBar: true,
-    pauseOnHover: true,
-    theme: 'light',
-    icon: false,
-};
-
-export enum TOAST_TYPE {
-    SUCCESS = 'success',
-    INFO = 'info',
-    ERROR = 'error',
-    WARNING = 'warning',
-    LOADING = 'loading',
-}
-
-export const DEFAULT_TOAST_MESSAGES = {
-    success: {
-        status: TOAST_TYPE.SUCCESS,
-        title: 'Action succeeded!',
-        description: 'The specified action went through!',
-    },
-    error: {
-        status: TOAST_TYPE.ERROR,
-        title: 'Something went wrong!',
-        description: 'An unexpected error occurred. Please try again later.',
-    },
-};
